Fix bug in evaluator for iter-Nat
Previously, iter-Nat added the same type annotations to its neutral
form as which-Nat, and our tests didn't catch it because they only
applied iter-Nat to computing Nats. This is probably a copy-paste bug.

Now, iter-Nat uses the right type on its neutral form, and a test has
been added.

Fixes #4.